src/backend/requirements.txt
django>=5.2.7
django-cors-headers>=4.9.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
django-filter>=25.2
pillow>=12.0.0
psycopg2-binary>=2.9.11
python-decouple>=3.8
gunicorn>=23.0.0
whitenoise>=6.11.0
